server/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
google-genai==1.67.0
google-adk==0.1.0
Pillow==10.1.0
python-multipart==0.0.6
websockets==12.0
